requirements.txt
# CSVファイルビューアーアプリケーションの依存パッケージ
# 作成日: 2024年

# Webアプリケーションフレームワーク
streamlit==1.45.1

# データ処理
pandas==2.2.3

# 日本語テキスト処理
mecab-python3==1.0.10  # 形態素解析
neologdn==0.5.4        # 日本語テキスト正規化
demoji==1.1.0          # 絵文字処理

# インストール方法:
# pip install -r requirements.txt 